Refactor to DRY up methods for scrubbing sponsor id
